web3-eth-iban - move unit tests inside package
Unit tests can be ran from monorepo root with:
```
lerna run --scope web3-eth-iban test:unit
lerna run --scope web3-eth-iban test:unit -- test/iban.createIndirect.js
```
