+++ packages/web3-eth-iban/test/iban.toAddress.js
@@ -1,8 +1,8 @@
-var chai = require('chai');
-var assert = chai.assert;
-var Iban = require('../packages/web3-eth-iban');
+import chai from 'chai';
+import Iban from '../';
+const assert = chai.assert;
 
-var tests = [
+const tests = [
     { direct: 'XE7338O073KYGTWWZN0F2WZ0R8PX5ZPPZS', address: '0x00c5496aEe77C1bA1f0854206A26DdA82a81D6D8'}
 ];
